+++ e2e/cli/e2e-pnpm6/dist-tags.spec.ts
@@ -1,6 +1,22 @@
 import { addRegistry, initialSetup, prepareGenericEmptyProject } from '@verdaccio/test-cli-commons';
 
-import { bumbUp, pnpm, publish } from './utils';
+import { pnpm } from './utils';
+
+async function bumbUp(tempFolder, registry) {
+  await pnpm({ cwd: tempFolder }, 'version', 'minor', ...addRegistry(registry.getRegistryUrl()));
+}
+
+async function publish(tempFolder, pkgName, registry, arg: string[] = []) {
+  const resp = await pnpm(
+    { cwd: tempFolder },
+    'publish',
+    ...arg,
+    '--json',
+    ...addRegistry(registry.getRegistryUrl())
+  );
+  const parsedBody = JSON.parse(resp.stdout as string);
+  expect(parsedBody.name).toEqual(pkgName);
+}
 
 describe('publish a package', () => {
   jest.setTimeout(20000);
